#167 Moved the class UriInfoServiceUrlProvider to io.katharsis.rs.resource.registry package and Exported 'io.katharsis.queryspec.*' in katharsis-core
