+++ katharsis-rs/src/main/java/io/katharsis/rs/KatharsisFeature.java
@@ -32,7 +32,7 @@ import io.katharsis.resource.registry.ResourceLookup;
 import io.katharsis.resource.registry.ResourceRegistry;
 import io.katharsis.resource.registry.ResourceRegistryBuilder;
 import io.katharsis.resource.registry.ServiceUrlProvider;
-import io.katharsis.resource.registry.UriInfoServiceUrlProvider;
+import io.katharsis.rs.resource.registry.UriInfoServiceUrlProvider;
 import io.katharsis.rs.parameterProvider.RequestContextParameterProviderLookup;
 import io.katharsis.rs.parameterProvider.RequestContextParameterProviderRegistry;
 import io.katharsis.rs.parameterProvider.RequestContextParameterProviderRegistryBuilder;
